requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
greenlet==3.1.1
idna==3.10
numpy==2.0.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
